relationship.select_one() allows to keep missing relationship
the idea is to allow to keep the rows when an agent tries to select a SIM from a dealer which is out of stock (i.e. is no longer in the 'from' part of the relationship). => in such case, we want to be able to write a scenario that keep the rows s.t. we can force that agent to try again at next clock stop