bugfix: collisions between reset_timer and force_act_next
previously, if an action was executing after a call to force_act_next, and if some actors had been marked as forced, the call to reset_timer was cancelling that mark. This is now fixed. This is important for the 'stock level influences sales' use case, since in that case we want some actor to force themselves to act next, in order to retry at next clock tick